+++ v2/pkg/engine/resolve/resolve.go
@@ -165,6 +165,8 @@ type ResolverOptions struct {
 	// upstream subgraph requests. This extension explains why each field was requested.
 	// This flag does not expose the data to clients.
 	PropagateFetchReasons bool
+
+	ValidateRequiredExternalFields bool
 }
 
 // New returns a new Resolver. ctx.Done() is used to cancel all active subscriptions and streams.
@@ -248,6 +250,7 @@ func newTools(options ResolverOptions, allowedExtensionFields map[string]struct{
 			allowAllErrorExtensionFields:                 options.AllowAllErrorExtensionFields,
 			apolloRouterCompatibilitySubrequestHTTPError: options.ApolloRouterCompatibilitySubrequestHTTPError,
 			propagateFetchReasons:                        options.PropagateFetchReasons,
+			validateRequiredExternalFields:               options.ValidateRequiredExternalFields,
 		},
 	}
 }
@@ -954,6 +957,18 @@ func (r *Resolver) AsyncUnsubscribeSubscription(id SubscriptionIdentifier) error
 		id:   id,
 		kind: subscriptionEventKindRemoveSubscription,
 	}:
+	default:
+		// In the event we cannot insert immediately, defer insertion a goroutine, this should prevent deadlocks, at the cost of goroutine creation.
+		go func() {
+			select {
+			case <-r.ctx.Done():
+				return
+			case r.events <- subscriptionEvent{
+				id:   id,
+				kind: subscriptionEventKindRemoveSubscription,
+			}:
+			}
+		}()
 	}
 	return nil
 }
@@ -968,6 +983,20 @@ func (r *Resolver) AsyncUnsubscribeClient(connectionID int64) error {
 		},
 		kind: subscriptionEventKindRemoveClient,
 	}:
+	default:
+		// In the event we cannot insert immediately, defer insertion a goroutine, this should prevent deadlocks, at the cost of goroutine creation.
+		go func() {
+			select {
+			case <-r.ctx.Done():
+				return
+			case r.events <- subscriptionEvent{
+				id: SubscriptionIdentifier{
+					ConnectionID: connectionID,
+				},
+				kind: subscriptionEventKindRemoveClient,
+			}:
+			}
+		}()
 	}
 	return nil
 }
@@ -1126,7 +1155,14 @@ func (r *Resolver) AsyncResolveGraphQLSubscription(ctx *Context, subscription *G
 		return writeFlushComplete(writer, msg)
 	}
 
-	event := subscriptionEvent{
+	select {
+	case <-r.ctx.Done():
+		// Stop resolving if the resolver is shutting down
+		return r.ctx.Err()
+	case <-ctx.ctx.Done():
+		// Stop resolving if the client is gone
+		return ctx.ctx.Err()
+	case r.events <- subscriptionEvent{
 		triggerID: xxh.Sum64(),
 		kind:      subscriptionEventKindAddSubscription,
 		addSubscription: &addSubscription{
@@ -1137,13 +1173,7 @@ func (r *Resolver) AsyncResolveGraphQLSubscription(ctx *Context, subscription *G
 			id:        id,
 			completed: make(chan struct{}),
 		},
-	}
-
-	select {
-	case <-r.ctx.Done():
-		// Stop resolving if the resolver is shutting down
-		return r.ctx.Err()
-	case r.events <- event:
+	}:
 	}
 	return nil
 }